feat(opensphere): improve tests
only relying on internal configs to run tests; while allowing full configurability for dry runs, etc
