--- .jscodeshift.json
@@ -1,10 +1,0 @@
-{
-  "globals": {
-    "os.alertManager": {"require": "os.alert.AlertManager", "singleton": true},
-    "os.feature": {},
-    "os.geo": {},
-    "os.ui": {},
-    "os.settings": {"require": "os.config.Settings", "singleton": true},
-    "os.MapContainer": {}
-  }
-}
